Make js functions for the buttons that change the shopping cart
Make functions in js that call actions from the UserCartController to make changes to the cart, such as adding or removing products. The UserCartController returns a view component of the cart to the JavaScript, as HTML text. This HTML is used to change the page, updating the cart, without reloading the page.